+++ controllers/product/CRUD/update.js
@@ -34,7 +34,7 @@ const updateProduct = async (req, res) => {
             if (result.value.coverImage) {
                 await storage.bucket(`${process.env.GCS_BUCKET}`).file(result.value.coverImage.name).delete();
             }
-            if (result.value.gallery && result.value.gallery.length > 0) {
+            if (files[1] && result.value.gallery && result.value.gallery.length > 0) {
                 result.value.gallery.map(async (image) => await storage.bucket(`${process.env.GCS_BUCKET}`).file(image.name).delete());
             }
         }
